+++ src/Accordion.php
@@ -96,6 +96,7 @@ class Accordion extends Widget
      * accordion effect. Set this to `false` to allow keeping multiple items open at once.
      */
     public $autoCloseItems = true;
+
     /**
      * @var array the HTML options for the item toggle tag. Key 'tag' might be used here for the tag name specification.
      * For example:
@@ -110,6 +111,19 @@ class Accordion extends Widget
      */
     public $itemToggleOptions = [];
 
+    /**
+     * @var array the HTML options for the header toggle tag. Key 'tag' might be used here for the tag name specification.
+     * Merged to:
+     *
+     * ```php
+     * [
+     *     'tag' => 'h5',
+     *     'class' => 'mb-0',
+     * ]
+     * ```
+     *
+     */
+    public $headerToggleOptions = [];
 
     /**
      * {@inheritdoc}
@@ -214,7 +228,13 @@ class Accordion extends Widget
                     ]) . "\n";
             }
 
-            $header = Html::tag('h5', $headerToggle, ['class' => 'mb-0']);
+            $headerToggleOptions = array_merge([
+                'tag' => 'h5',
+                'class' => 'mb-0',
+            ], $this->headerToggleOptions);
+
+            $headerToggleTag = ArrayHelper::remove($headerToggleOptions, 'tag', 'h5');
+            $header = Html::tag($headerToggleTag, $headerToggle, $headerToggleOptions);
 
             if (is_string($item['content']) || is_numeric($item['content']) || is_object($item['content'])) {
                 $content = Html::tag('div', $item['content'], ['class' => 'card-body']) . "\n";